+++ packages/twenty-server/src/engine/metadata-modules/flat-search-field-metadata/utils/derive-search-vector-as-expression-for-ts-vector-field.util.ts
@@ -1,5 +1,3 @@
-import { Logger } from '@nestjs/common';
-
 import { type FieldMetadataType } from 'twenty-shared/types';
 import { isDefined } from 'twenty-shared/utils';
 
@@ -9,19 +7,15 @@ import {
 } from 'src/engine/metadata-modules/flat-search-field-metadata/utils/compute-search-vector-as-expression-from-search-field-metadatas.util';
 import { type FlatSearchFieldMetadata } from 'src/engine/metadata-modules/flat-search-field-metadata/types/flat-search-field-metadata.type';
 
-const logger = new Logger('deriveSearchVectorAsExpressionForTsVectorField');
-
 export const deriveSearchVectorAsExpressionForTsVectorField = ({
   targetSearchFieldMetadatas,
   indexedFieldById,
-  objectMetadataNameSingular,
 }: {
   targetSearchFieldMetadatas: FlatSearchFieldMetadata[];
   indexedFieldById: ReadonlyMap<
     string,
     { name: string; type: FieldMetadataType }
   >;
-  objectMetadataNameSingular?: string;
 }): string => {
   const targetSearchableFields = targetSearchFieldMetadatas.flatMap(
     (flatSearchFieldMetadata) => {
@@ -42,16 +36,6 @@ export const deriveSearchVectorAsExpressionForTsVectorField = ({
       ];
     },
   );
-
-  if (targetSearchableFields.length === 0) {
-    logger.error(
-      `Deriving searchVector expression with zero target search fields${
-        isDefined(objectMetadataNameSingular)
-          ? ` for object "${objectMetadataNameSingular}"`
-          : ''
-      }; the searchVector column will be NULL and full-text search will not work.`,
-    );
-  }
 
   return computeSearchVectorAsExpressionFromSearchFieldMetadatas(
     targetSearchableFields,
