+++ packages/twenty-server/src/engine/metadata-modules/object-metadata/utils/build-default-relation-flat-field-metadatas-for-custom-object.util.ts
@@ -167,11 +167,11 @@ export const buildDefaultRelationFlatFieldMetadatasForCustomObject = ({
         return {
           standardSourceFlatFieldMetadatas: [
             ...sourceAndTargetFlatFieldMetadatasRecord.standardSourceFlatFieldMetadatas,
-            { ...flatFieldMetadatas[0], isSystemSideEffect: false },
+            flatFieldMetadatas[0],
           ],
           standardTargetFlatFieldMetadatas: [
             ...sourceAndTargetFlatFieldMetadatasRecord.standardTargetFlatFieldMetadatas,
-            { ...flatFieldMetadatas[1], isSystemSideEffect: false },
+            flatFieldMetadatas[1],
           ],
           standardTargetFlatIndexMetadatas: [
             ...sourceAndTargetFlatFieldMetadatasRecord.standardTargetFlatIndexMetadatas,
