+++ packages/twenty-server/src/engine/metadata-modules/object-metadata/utils/build-reserved-system-flat-field-metadatas-for-custom-object.util.ts
@@ -1,11 +1,10 @@
 import { getFieldUniversalIdentifier } from 'twenty-shared/application';
-import { FieldMetadataType } from 'twenty-shared/types';
 
 import { PARTIAL_SYSTEM_FLAT_FIELD_METADATAS } from 'src/engine/metadata-modules/object-metadata/constants/partial-system-flat-field-metadatas.constant';
 import { type UniversalFlatFieldMetadata } from 'src/engine/workspace-manager/workspace-migration/universal-flat-entity/types/universal-flat-field-metadata.type';
 import { type UniversalFlatObjectMetadata } from 'src/engine/workspace-manager/workspace-migration/universal-flat-entity/types/universal-flat-object-metadata.type';
 
-type BuildFlatFieldMetadataForCustomObjectArgs = {
+type BuildReservedSystemFlatFieldMetadatasForCustomObjectArgs = {
   flatObjectMetadata: Pick<
     UniversalFlatObjectMetadata,
     'universalIdentifier' | 'applicationUniversalIdentifier'
@@ -17,7 +16,7 @@ export const buildReservedSystemFlatFieldMetadatasForCustomObject = ({
     applicationUniversalIdentifier,
     universalIdentifier: objectMetadataUniversalIdentifier,
   },
-}: BuildFlatFieldMetadataForCustomObjectArgs): Record<
+}: BuildReservedSystemFlatFieldMetadatasForCustomObjectArgs): Record<
   string,
   UniversalFlatFieldMetadata
 > => {
@@ -98,77 +97,4 @@ export const buildReservedSystemFlatFieldMetadatasForCustomObject = ({
       updatedAt: now,
     },
   } as const satisfies Record<string, UniversalFlatFieldMetadata>;
-};
-
-export const buildSearchVectorFlatFieldMetadataForCustomObject = ({
-  flatObjectMetadata: {
-    applicationUniversalIdentifier,
-    universalIdentifier: objectMetadataUniversalIdentifier,
-  },
-}: BuildFlatFieldMetadataForCustomObjectArgs): UniversalFlatFieldMetadata<FieldMetadataType.TS_VECTOR> => {
-  const now = new Date().toISOString();
-
-  const { searchVector } = PARTIAL_SYSTEM_FLAT_FIELD_METADATAS;
-
-  return {
-    ...searchVector,
-    universalIdentifier: getFieldUniversalIdentifier({
-      applicationUniversalIdentifier,
-      objectUniversalIdentifier: objectMetadataUniversalIdentifier,
-      name: searchVector.name,
-    }),
-    applicationUniversalIdentifier,
-    objectMetadataUniversalIdentifier,
-    createdAt: now,
-    updatedAt: now,
-    universalSettings: null,
-  };
-};
-
-export const buildNameFlatFieldMetadataForCustomObject = ({
-  flatObjectMetadata: {
-    applicationUniversalIdentifier,
-    universalIdentifier: objectMetadataUniversalIdentifier,
-  },
-}: BuildFlatFieldMetadataForCustomObjectArgs): UniversalFlatFieldMetadata<FieldMetadataType.TEXT> => {
-  const now = new Date().toISOString();
-
-  return {
-    type: FieldMetadataType.TEXT,
-    isLabelSyncedWithName: false,
-    isUnique: false,
-    universalIdentifier: getFieldUniversalIdentifier({
-      applicationUniversalIdentifier,
-      objectUniversalIdentifier: objectMetadataUniversalIdentifier,
-      name: 'name',
-    }),
-    name: 'name',
-    label: 'Name',
-    icon: 'IconAbc',
-    description: 'Name',
-    isNullable: true,
-    isActive: true,
-    isSystem: false,
-    isSystemSideEffect: false,
-    isUIEditable: true,
-    defaultValue: null,
-    createdAt: now,
-    updatedAt: now,
-    options: null,
-    overrides: null,
-    morphId: null,
-    applicationUniversalIdentifier,
-    objectMetadataUniversalIdentifier,
-    relationTargetObjectMetadataUniversalIdentifier: null,
-    relationTargetFieldMetadataUniversalIdentifier: null,
-    viewFilterUniversalIdentifiers: [],
-    viewFieldUniversalIdentifiers: [],
-    kanbanAggregateOperationViewUniversalIdentifiers: [],
-    calendarViewUniversalIdentifiers: [],
-    mainGroupByFieldMetadataViewUniversalIdentifiers: [],
-    fieldPermissionUniversalIdentifiers: [],
-    universalSettings: null,
-    viewSortUniversalIdentifiers: [],
-    searchFieldMetadataUniversalIdentifiers: [],
-  };
 };
