+++ packages/twenty-server/src/engine/workspace-manager/workspace-migration/workspace-migration-runner/action-handlers/field/services/update-field-action-handler.service.ts
@@ -378,7 +378,6 @@ export class UpdateFieldActionHandlerService extends WorkspaceMigrationRunnerAct
               flatSearchFieldMetadataMaps,
             }),
           indexedFieldById,
-          objectMetadataNameSingular: flatObjectMetadata.nameSingular,
         });
 
       const columnDefinitions = generateColumnDefinitions({
