+++ packages/twenty-server/src/engine/workspace-manager/workspace-migration/workspace-migration-runner/action-handlers/object/services/create-object-action-handler.service.ts
@@ -162,7 +162,6 @@ export class CreateObjectActionHandlerService extends WorkspaceMigrationRunnerAc
                     allFlatEntityMaps.flatSearchFieldMetadataMaps,
                 }),
               indexedFieldById,
-              objectMetadataNameSingular: flatObjectMetadata.nameSingular,
             })
           : undefined,
       }),
